move tests of model to tests/model
