+++ tests/models/conftest.py
@@ -1,0 +1,45 @@
+# Copyright (c) 2024, RTE (https://www.rte-france.com)
+#
+# See AUTHORS.txt
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+# SPDX-License-Identifier: MPL-2.0
+#
+# This file is part of the Antares project.
+from pathlib import Path
+
+import pytest
+
+from andromede.model.parsing import parse_yaml_library
+from andromede.model.resolve_library import resolve_library
+
+
+@pytest.fixture(scope="session")
+def data_dir() -> Path:
+    return Path(__file__).parents[1] / "unittests/data"
+
+
+@pytest.fixture(scope="session")
+def lib(data_dir: Path):
+    lib_file = data_dir / "lib.yml"
+
+    with lib_file.open() as f:
+        input_lib = parse_yaml_library(f)
+
+    lib = resolve_library(input_lib)
+    return lib
+
+
+@pytest.fixture(scope="session")
+def lib_sc(data_dir: Path):
+    libs_path = Path(__file__).parents[2] / "src/andromede/libs/"
+    lib_sc_file = libs_path / "standard_sc.yml"
+
+    with lib_sc_file.open() as f:
+        input_lib_sc = parse_yaml_library(f)
+
+    lib_sc = resolve_library(input_lib_sc)
+    return lib_sc
