+++ tests/models/test_electrolyzer_n_inputs_yaml.py
