+++ tests/unittests/conftest.py
@@ -13,33 +13,7 @@ from pathlib import Path
 
 import pytest
 
-from andromede.model.parsing import parse_yaml_library
-from andromede.model.resolve_library import resolve_library
-
 
 @pytest.fixture(scope="session")
 def data_dir() -> Path:
     return Path(__file__).parent / "data"
-
-
-@pytest.fixture(scope="session")
-def lib(data_dir: Path):
-    lib_file = data_dir / "lib.yml"
-
-    with lib_file.open() as f:
-        input_lib = parse_yaml_library(f)
-
-    lib = resolve_library(input_lib)
-    return lib
-
-
-@pytest.fixture(scope="session")
-def lib_sc(data_dir: Path):
-    libs_path = Path(__file__).parents[2] / "src/andromede/libs/"
-    lib_sc_file = libs_path / "standard_sc.yml"
-
-    with lib_sc_file.open() as f:
-        input_lib_sc = parse_yaml_library(f)
-
-    lib_sc = resolve_library(input_lib_sc)
-    return lib_sc
